fuzzy tests extracted
